Improve speed of PIN validation
PIN validation on windows needs to be faster than 750ms.
The decryption operation of the user private key over HPKE is pretty slow

This PR moves this operation to future steps that need access to the key.
It's mostly moving the cost to later, non time-sensitive operations, but it also
saves computation for operations that need to be authenticated but don't need access to the key